+++ CppSource/LM75/lm75.cpp
@@ -23,7 +23,7 @@ uint16_t LM75::readTemperature() {
 		if (noisyTemperature >= 125.0) {
 			rawData = 250 << 7;
 		} else {
-			tempValue = noisyTemperature / 0.5;
+			tempValue = int(noisyTemperature / 0.5);
 			rawData = tempValue << 7;
 		}
 	} else { // Negative temperatures
@@ -31,7 +31,7 @@ uint16_t LM75::readTemperature() {
 			rawData = 1 << 15;
 			rawData = rawData | 146 << 7;
 		} else {
-			tempValue = 256 + (noisyTemperature / 0.5);
+			tempValue = int(256 + (noisyTemperature / 0.5));
 			rawData = 1 << 15;
 			rawData = rawData | tempValue << 7;
 		}
